# openapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Reservation System API",
    "version": "1.0.0",
    "description": "Reservation System API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "Swagger API Team",
      "url": "http://swagger.io",
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/businesses": {
      "get": {
        "operationId": "getBusinesses",
        "summary": "List all businesses",
        "responses": {
          "200": { "description": "List of businesses" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "operationId": "postBusinesses",
        "summary": "Create a new business",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE", "SUSPENDED"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Business created" },
          "400": { "description": "Invalid payload" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/businesses/{business_id}": {
      "parameters": [
        { "name": "business_id", "in": "path", "required": true, "schema": { "type": "integer" } }
      ],
      "get": {
        "operationId": "getBusinessesBusinessId",
        "summary": "Get a business by ID",
        "responses": {
          "200": { "description": "Business found" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Business not found" }
        }
      },
      "put": {
        "operationId": "putBusinessesBusinessId",
        "summary": "Update a business",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "status": { "type": "string", "enum": ["ACTIVE", "INACTIVE", "SUSPENDED"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Business updated" },
          "400": { "description": "Invalid payload" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Business not found" }
        }
      },
      "delete": {
        "operationId": "deleteBusinessesBusinessId",
        "summary": "Delete a business",
        "responses": {
          "200": { "description": "Business deleted" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Business not found" }
        }
      }
    },
    "/businesses/{business_id}/locations": {
      "parameters": [
        { "name": "business_id", "in": "path", "required": true, "schema": { "type": "integer" } }
      ],
      "get": {
        "operationId": "getBusinessesBusinessIdLocations",
        "summary": "List all locations for a business",
        "responses": {
          "200": { "description": "List of locations" },
          "400": { "description": "Invalid business ID" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/locations": {
      "get": {
        "operationId": "getLocations",
        "summary": "List all locations",
        "parameters": [
          { "name": "business_id", "in": "query", "required": false, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "List of locations" },
          "400": { "description": "Invalid filters" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "operationId": "postLocations",
        "summary": "Create a new location",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "city", "postalCode", "address", "businessId"],
                "properties": {
                  "name": { "type": "string" },
                  "city": { "type": "string" },
                  "postalCode": { "type": "string" },
                  "address": { "type": "string" },
                  "businessId": { "type": "integer" },
                  "latitude": { "type": "number", "nullable": true },
                  "longitude": { "type": "number", "nullable": true },
                  "phone": { "type": "string", "nullable": true },
                  "email": { "type": "string", "format": "email", "nullable": true },
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Location created" },
          "400": { "description": "Invalid payload" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/locations/{location_id}": {
      "parameters": [
        { "name": "location_id", "in": "path", "required": true, "schema": { "type": "integer" } }
      ],
      "get": {
        "operationId": "getLocationsLocationId",
        "summary": "Get a location by ID",
        "responses": {
          "200": { "description": "Location found" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Location not found" }
        }
      },
      "put": {
        "operationId": "putLocationsLocationId",
        "summary": "Update a location",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "city": { "type": "string" },
                  "postalCode": { "type": "string" },
                  "address": { "type": "string" },
                  "businessId": { "type": "integer" },
                  "latitude": { "type": "number", "nullable": true },
                  "longitude": { "type": "number", "nullable": true },
                  "phone": { "type": "string", "nullable": true },
                  "email": { "type": "string", "format": "email", "nullable": true },
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Location updated" },
          "400": { "description": "Invalid payload" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Location not found" }
        }
      },
      "delete": {
        "operationId": "deleteLocationsLocationId",
        "summary": "Delete a location",
        "responses": {
          "200": { "description": "Location deleted" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Location not found" }
        }
      }
    }
  },
...
}